refactoring, extract SNI
